csd-background: Make wayland-compatible.
This may not last, as it's not able to actually run as a wayland
client currently (due to the fact we can't hint to muffin that
these are desktop/background windows.

New code is added in muffin to sort these windows to the bottom
of the bottom_window_group (behind file managers).

Behave as before under x11 (painting the root window).